Safari-611.1.10/LayoutTests/svg/custom/svg-createsvgtransform-type.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Tests whether createSVGTransform creates a SVGTransform of the right SVGTransform type.");

var svgDoc = document.implementation.createDocument("http://www.w3.org/2000/svg", "svg", null);
var transform = svgDoc.documentElement.createSVGTransform();
shouldBe("transform.type", "SVGTransform.SVG_TRANSFORM_MATRIX");

var successfullyParsed = true;
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
